Wallet to wallet app. Nextjs latest structure used. solana-pay.md
I changed places referrering to wallet softwares from wallet to wallet app. 

Pages/api has become deprecated in the latest Nextjs version so I introduced the latest structure for the intro sections of the app. Will be updating the scavenger code to use the latest nextjs structure before completing the codes in this course.

I however indicated this to the reader in the course material