fix and improve close-mint course (#536)
* Update the Token Extensions Program link

* sync token-program.md from main

* change "mint" to "token mint account".
add a typescript code example for using an account on devnet while calling "initializeKeypair" in the section [2. Run validator node]
add a typescript code example for importing our new function in the section [3. Create a mint with close authority]
change "const mintInfo" to "let mintInfo" because "mintInfo" is used again in the "CLOSE MINT" code part.

---------

Co-authored-by: Jeff Wood <wujunadmn@gmail.com>